[Core] - Implement A Few Things
* Implemented aura stacks functionality
* Implemented new server opcode SMSG_AURA_UPDATE
* Fixed a bug with AoE spells
* Implemented UIHandler.cs - Will hold all of the necessary functionality to run the UI. I'll still probably use a separate script for each UI object but at least I can control them all from one location.

+++ Assets/Spell/Auras/AuraInfo.cs
@@ -17,7 +17,7 @@ public class AuraInfo
 
     public float TickTime { get; private set; }
 
-    public float Stacks { get; private set; }
+    public int Stacks { get; private set; }
 
     public int BasePoints { get; private set; }
